# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Swadista Organic Store",
    "description": "This is a sample API for Oraganic store.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/users/signup": {
      "post": {
        "tags": ["Signup"],
        "summary": "Create a new user",
        "description": "Create a new user with email and phone number.",

        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/signup"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/signup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/signup"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/signup"
                }
              }
            }
          },
          "400": {
            "description": "Email or phone number already exists"
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "tags": ["Login"],
        "summary": "Login a user",
        "description": "Login user with email and phone number.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/login"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/login"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/login"
                }
              }
            }
          },
          "400": {
            "description": "Email or phone number already exists"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "signup": {
        "type": "object",
        "required": [
          "email",
          "phone_number",
          "password",
          "first_name",
          "first_name"
        ],

        "properties": {
          "first_name": {
            "type": "string",
            "description": "The user's first name"
          },
          "last_name": {
            "type": "string",
            "description": "The user's last name"
          },
          "email": {
            "type": "string",
            "description": "The user's email address"
          },
          "phone_number": {
            "type": "string",
            "description": "The user's phone number"
          },
          "password": {
            "type": "string",
            "description": "The user's password"
          }
        }
      },
...
      "login": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "description": "The user's email address"
          },
          "phone_number": {
            "type": "string",
            "description": "The user's phone number"
          },
          "password": {
            "type": "string",
            "description": "The user's password"
          }
        }
      },
...
    }
  }
}